+++ zed_debug/launch/zed_camera_debug.launch.py
@@ -171,6 +171,7 @@ def launch_setup(context, *args, **kwargs):
         camera_model_val == 'zed2i' or
         camera_model_val == 'zedx' or
         camera_model_val == 'zedxm' or
+        camera_model_val == 'zedxnano' or
         camera_model_val == 'zedxhdr' or
         camera_model_val == 'zedxhdrmini' or
         camera_model_val == 'zedxhdrmax' or
@@ -332,7 +333,7 @@ def generate_launch_description():
             DeclareLaunchArgument(
                 'camera_model',
                 description='[REQUIRED] The model of the camera. Using a wrong camera model can disable camera features.',
-                choices=['zed', 'zedm', 'zed2', 'zed2i', 'zedx', 'zedxm', 'zedxhdr', 'zedxhdrmini', 'zedxhdrmax', 'virtual', 'zedxonegs', 'zedxone4k', 'zedxonehdr']),
+                choices=['zed', 'zedm', 'zed2', 'zed2i', 'zedx', 'zedxm', 'zedxnano', 'zedxhdr', 'zedxhdrmini', 'zedxhdrmax', 'virtual', 'zedxonegs', 'zedxone4k', 'zedxonehdr']),
             DeclareLaunchArgument(
                 'namespace',
                 default_value='',